Separate HotReload into a library&executable QaterialHotReloadLib & QaterialHotReload
