+++ tools/CMakeLists.txt
@@ -1,2 +1,8 @@
-add_subdirectory(HotReload)
-add_subdirectory(HelloWorld)
+
+if(QATERIAL_ENABLE_HOTRELOAD OR QATERIAL_ENABLE_HOTRELOAD_LIB)
+  add_subdirectory(HotReload)
+endif()
+
+if(QATERIAL_ENABLE_HELLOWORLD)
+  add_subdirectory(HelloWorld)
+endif()
