+++ tools/HotReload/qml/IconsMenu.qml
