+++ tools/HotReload/qml/QaterialHotReload.qrc
