+++ tools/HotReload/qml/StatusView.qml
