+++ tools/HotReload/src/HotReload.cpp
@@ -1,0 +1,25 @@
+// ──── INCLUDE ────
+
+// Oliv Qt
+#include <Qaterial/HotReload/HotReload.hpp>
+
+// ──── DECLARATION ────
+
+using namespace qaterial;
+
+const std::shared_ptr<HotReloadSink> HotReload::_sink = std::make_shared<HotReloadSink>();
+
+void HotReloadSink::sink_it_(const spdlog::details::log_msg& msg)
+{
+    if(!_hotReload)
+        return;
+
+    // log_msg is a struct containing the log entry info like level, timestamp, thread id etc.
+    // msg.raw contains pre formatted log
+
+    // If needed (very likely but not mandatory), the sink formats the message before sending it to its final destination:
+    spdlog::memory_buf_t formatted;
+    base_sink<std::mutex>::formatter_->format(msg, formatted);
+    const auto string = fmt::to_string(formatted);
+    Q_EMIT _hotReload->newLog(QString::fromStdString(string));
+}
